[move] IR module definitions require an address
In Move source language, module definitions must either have leading
addresses specified (`module 0x42::M { ... }`), or be defined within
an address definition (`address 0x42 { module M { ... } }`). In Move IR,
however, modules cannot be defined with an address, and instead
command-line tools that process the IR specify the address under which
to publish the module.

Remove the ability to pass in addresses from the command-line with Move
IR, and instead require modules to be defined with a leading address.