[move] Expand the characters permitted in comments
Within comments that are not exported as documentation, allow for the
use of nearly all code points, with the exception of specific whitespace
characters that may cause confusion depending on the text renderer being
used.

See the newly added tests for details on what characters are allowed in
which contexts. Basically, the rules could be summarized as follows:

* Within single-line, non-documentation comments: all characters are
  permitted, except for vertical whitespace/line-break characters other
  than `\n` (`U+000A LINE FEED`).
* Within multi-line, non-documentation comments: all characters are
  permitted, except for horizontal whitespace characters other than ` `
  and `\t` (`U+0020 SPACE` and `U+0009 CHARACTER TABULATION`), and
  except for vertical whitespace characters other than `\n`.
* Within single-line and multi-line documentation comments: only
  characters permitted in un-commented Move source code is permitted --
  all of these characters are part of the ASCII character set, with some
  esoteric code points excluded (such as `U+000B LINE TABULATION` and
  `U+000D CARRIAGE RETURN`).

The intent is to allow Move users to write comments in their preferred
language, but disallow characters that make it ambiguous as to where the
commented-out portion of the source code ends.