[WIP][move] Add failing test for toycoin
When following the instructions in the Move book to create a toy coin
module and script, the first script example fails to compile:

```
thread 'run_test::creating-coins.move' panicked at 'called `Option::unwrap()` on a `None` value', language/move-lang/src/to_bytecode/translate.rs:1005:77
```

Currently I don't believe we have tests for the examples in the Move
book. Add a test harness for these, as well as a failing test for the
toy coin example.