[move] Migrate VMControlFlowGraph test
The only tests for VMControlFlowGraph itself exist in the
move-ir-compiler unit tests, but that's not an ideal location. Instead,
move the tests closer to where the type is defined: the
move-binary-format crate.

However, as-is the tests are written using Move IR strings that are
translated into Move bytecode, which is used to construct the CFG. Since
the ir-to-bytecode routines use types defined in move-binary-format, a
dev dependency cycle would be introduced by adding these tests as
currently written to the move-binary-format unit tests (the tests of the
nove-binary-format crate would depend on move-ir-to-bytecode, which
depend on move-binary-format). Instead, define them as integration tests
(which depend on both move-binary-format and move-ir-to-bytecode as
external crate dependencies).